fixed: publish service in test network with different configuration
