+++ registry/testcases/unit_testcases/test_organization_publisher_service.py
@@ -109,8 +109,8 @@ class TestOrganizationPublisherService(unittest.TestCase):
         assert True
 
     @patch("common.ipfs_util.IPFSUtil", return_value=Mock(write_file_in_ipfs=Mock(return_value="Q12PWP")))
-    @patch("registry.domain.services.organization_domain_service."
-           "OrganizationService.publish_organization_to_test_network", return_value="0x123")
+    @patch("registry.domain.services.registry_blockchain_util."
+           "RegistryBlockChainUtil.publish_organization_to_test_network", return_value="0x123")
     def test_org_publish_to_ipfs(self, mock_test_network_publish, mock_ipfs_utils):
         test_org_id = uuid4().hex
         username = "dummy@snet.io"
